I break down monolith components into sub-components for more modular architecture.
