+++ src/components/FormInput/index.js
@@ -1,0 +1,3 @@
+import FormInput from './FormInput'
+
+export default FormInput
