+++ src/components/MintForm/index.js
@@ -1,0 +1,3 @@
+import MintForm from './MintForm'
+
+export default MintForm
