+++ src/components/MintFormFields/index.js
@@ -1,0 +1,3 @@
+import MintFormFields from './MintFormFields'
+
+export default MintFormFields
